:truck: add EmbeddedMongo profile for testing /rename all IT classes to Test classes: give up on using failsafe for the moment
