+++ Spring-Boot/Spring-Boot-Commons/src/main/resources/config/.gitignore
@@ -1,2 +1,2 @@
 application.yml
-../application.properties
+application-mongo.yml
